feat: add autoEllipsis method to legend
